+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,51 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class AppUserService {
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+    private final DataSource dataSource;
+
+    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final DataSource dataSource) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+        this.dataSource = dataSource;
+    }
+
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    public void insert(final User user) {
+        userDao.insert(user);
+    }
+
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        try (final Connection connection = dataSource.getConnection()) {
+            connection.setAutoCommit(false);
+
+            try {
+                final User user = userDao.transactionFindById(connection, id);
+                user.changePassword(newPassword);
+                userDao.transactionUpdate(connection, user);
+                userHistoryDao.log(connection, new UserHistory(user, createBy));
+
+                connection.commit();
+            } catch (Exception e) {
+                connection.rollback();
+                throw e;
+            }
+        } catch (SQLException e) {
+            throw new DataAccessException(e);
+        }
+    }
+}
